api: Add extra adapter pool-tagging should the "friendly name" is reset
On Windows 10 major release upgrades the network adapters were recreated
and the friendly name we set to our Wintun adapter for nicer display and
(important!) for identifying pool member adapters were reset to ""Wintun
Userspace Tunnel".

This commit adds extra tagging into the "DevRegKey" of the adapter.

However, the testing with Windows 10 2004 upgrade revealed that the
upgrade process resets the adapter list and Wintun adapters (presumably
having no physical hardware device) are lost. This means the extra
tagging is excessive, making this commit a flop.

Signed-off-by: Simon Rozman <simon@rozman.si>